requirements.txt
# DFTE — Dual Field Trading Engine Dependencies

# Market data
yfinance>=0.2.0

# API + server
fastapi>=0.100.0
uvicorn>=0.20.0
httpx>=0.24.0
pydantic>=2.0.0
requests>=2.28.0

# Data
numpy>=1.24.0
google-cloud-storage>=2.0.0
ephem>=4.1

# Terminal dashboard
rich>=13.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Optional: Alpaca execution
# alpaca-py>=0.10.0
# Set: ALPACA_API_KEY, ALPACA_SECRET_KEY
# Set: ALPACA_LIVE=true for live trading (default: paper)

# Optional: FRED macro data
# Set: FRED_API_KEY (free at https://fred.stlouisfed.org)

# BMR server must be running separately (kindpath-bmr repo)
# Default: http://localhost:8001
# Override: BMR_SERVER=http://your-server:8001
